+++ src/colours.js
@@ -1,8 +1,7 @@
 var vscode = require( 'vscode' );
-
 var utils = require( './utils.js' );
 
-function validate( workspace )
+function validateColours( workspace )
 {
     function check( setting )
     {
@@ -33,4 +32,36 @@ function validate( workspace )
     return result;
 }
 
-module.exports.validate = validate;
+function validateIconColours( workspace )
+{
+    function checkIconColour( setting )
+    {
+        var icon = workspace.getConfiguration( 'todo-tree.highlights' ).get( setting + ".icon" );
+        var iconColour = workspace.getConfiguration( 'todo-tree.highlights' ).get( setting + ".iconColour" );
+        if( icon.indexOf( "$(" ) != 0 && utils.isThemeColour( iconColour ) )
+        {
+            invalidIconColours.push( setting + '.iconColour (' + iconColour + ')' );
+        }
+    }
+
+    var invalidIconColours = [];
+    var result = "";
+
+    checkIconColour( 'defaultHighlight' );
+
+    var config = vscode.workspace.getConfiguration( 'todo-tree.highlights' );
+    Object.keys( config.customHighlight ).forEach( function( tag )
+    {
+        checkIconColour( 'customHighlight.' + tag );
+    } );
+
+    if( invalidIconColours.length > 0 )
+    {
+        result = "Invalid icon colour settings (theme colours can only be used with codicons): " + invalidIconColours.join( ', ' );
+    }
+
+    return result;
+}
+
+module.exports.validateColours = validateColours;
+module.exports.validateIconColours = validateIconColours;