+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Todo Tree Change Log
 
 - Make storage folder creation (if required) more robust
+- Improve icon colouring under error conditions
 
 ## v0.0.193 - 2020-12-31